Dcs branch

+++ server/controllers/maintenanceController.js
@@ -1,3 +1,5 @@
+import supabase from '../config/db.js';
+
 export const getMaintenanceLog = async (req, res) => {
     const userId = req.params.id
 
@@ -12,34 +14,69 @@ export const getMaintenanceLog = async (req, res) => {
         }
     return res.status(200).json(data);
     } catch (err) {
-        return res.status(500).json({ error: 'Server error' });
+        return res.status(500).json({ err: 'Server error' });
     }
 }
 
-
-
 export const addMaintenanceItem = async (req, res) => {
-    // add item and returns the item
+    const { serviceDate, serviceItem, mileage, notes } = req.body;
+    try {
     const { data, error } = await supabase
     .from('maintenance_log')
-    .insert({ })
+    .insert({ date_of_service: serviceDate, service_item: serviceItem, mileage, notes })
     .select()
+    .eq('id', userId);
+
+    if (error) {
+        return res.status(400).json({ error: error.message });
+    }
+    return res.status(200).json(data);
+
+    } catch (err) {
+        return res.status(500).json({err: 'Server error'});
+    }
 }
 
 export const editMaintenanceItem = async (req, res) => {
-    // update and returns the item
+    const updates = req.body;
+    try {
     const { data, error } = await supabase
     .from('maintenance_log')
-    .update({  })
-    .eq('id', 1)
+    .update(updates)
+    .eq('id', req.params.itemId)
     .select()
+    .eq('id', userId);
+
+    if (error) {
+        return res.status(400).json({ error: error.message });
+    }
+
+    return res.status(200).json(data);
+
+    } catch (err) {
+        return res.status(500).json({err: 'Server error'});
+    }
+
 }
 
 export const deleteMaintenanceItem = async (req, res) => {
-    // delete and returns the item
+    const {id} = req.body;
+    try {
     const { data, error } = await supabase
     .from('maintenance_log')
     .delete()
-    .eq('id', 1)
+    .eq('id', id)
     .select()
+    .eq('id', userId);
+
+    if (error) {
+        return res.status(400).json({ error: error.message });
+    }
+
+    return res.status(200).json(data);
+
+    } catch (err) {
+        return res.status(500).json({err: 'Server error'});
+    }
+
 }
